+++ packages/docusaurus-theme-classic/src/types.d.ts
@@ -312,7 +312,6 @@ declare module '@theme/Navbar' {
 
 declare module '@theme/NavbarItem/DefaultNavbarItem' {
   import type {LinkProps} from '@docusaurus/Link';
-  import type {LinkLikeNavbarItemProps} from '@theme/NavbarItem';
 
   export type NavLinkProps = LinkProps & {
     readonly activeBasePath?: string;
@@ -323,8 +322,6 @@ declare module '@theme/NavbarItem/DefaultNavbarItem' {
   };
 
   export type DesktopOrMobileNavBarItemProps = NavLinkProps & {
-    readonly items?: readonly LinkLikeNavbarItemProps[];
-    readonly position?: 'left' | 'right';
     readonly isDropdownItem?: boolean;
     readonly className?: string;
   };
@@ -333,8 +330,29 @@ declare module '@theme/NavbarItem/DefaultNavbarItem' {
     readonly mobile?: boolean;
   };
 
+  const NavLink: (props: NavLinkProps) => JSX.Element;
+  export const NavLink;
+
   const DefaultNavbarItem: (props: Props) => JSX.Element;
   export default DefaultNavbarItem;
+}
+
+declare module '@theme/NavbarItem/DropdownNavbarItem' {
+  import type {NavLinkProps} from '@theme/NavbarItem/DefaultNavbarItem';
+  import type {LinkLikeNavbarItemProps} from '@theme/NavbarItem';
+
+  export type DesktopOrMobileNavBarItemProps = NavLinkProps & {
+    readonly position?: 'left' | 'right';
+    readonly items: readonly LinkLikeNavbarItemProps[];
+    readonly className?: string;
+  };
+
+  export type Props = DesktopOrMobileNavBarItemProps & {
+    readonly mobile?: boolean;
+  };
+
+  const DropdownNavbarItem: (props: Props) => JSX.Element;
+  export default DropdownNavbarItem;
 }
 
 declare module '@theme/NavbarItem/SearchNavbarItem' {
@@ -398,17 +416,19 @@ declare module '@theme/NavbarItem' {
   import type {Props as DefaultNavbarItemProps} from '@theme/NavbarItem/DefaultNavbarItem';
   import type {Props as DocNavbarItemProps} from '@theme/NavbarItem/DocNavbarItem';
   import type {Props as DocsVersionNavbarItemProps} from '@theme/NavbarItem/DocsVersionNavbarItem';
+  import type {Props as DropdownNavbarItemProps} from '@theme/NavbarItem/DropdownNavbarItemProps';
   import type {Props as DocsVersionDropdownNavbarItemProps} from '@theme/NavbarItem/DocsVersionDropdownNavbarItem';
   import type {Props as LocaleDropdownNavbarItemProps} from '@theme/NavbarItem/LocaleDropdownNavbarItem';
   import type {Props as SearchNavbarItemProps} from '@theme/NavbarItem/SearchNavbarItem';
 
   export type LinkLikeNavbarItemProps =
-    | ({readonly type?: 'default' | undefined} & DefaultNavbarItemProps)
+    | ({readonly type?: 'default'} & DefaultNavbarItemProps)
     | ({readonly type: 'doc'} & DocNavbarItemProps)
     | ({readonly type: 'docsVersion'} & DocsVersionNavbarItemProps);
 
   export type Props =
     | LinkLikeNavbarItemProps
+    | ({readonly type: 'dropdown'} & DropdownNavbarItemProps)
     | ({
         readonly type: 'docsVersionDropdown';
       } & DocsVersionDropdownNavbarItemProps)
@@ -418,6 +438,8 @@ declare module '@theme/NavbarItem' {
         readonly position?: 'left' | 'right';
         readonly onClick?: React.MouseEventHandler<HTMLDivElement>;
       } & SearchNavbarItemProps);
+
+  export type Types = Props['type'];
 
   const NavbarItem: (props: Props) => JSX.Element;
   export default NavbarItem;
